refactor(cli): split binary into modules for Codacy file NLOC (#38)
Move quantize-goz1 and artifacts handlers into sibling modules so each
file stays under Codacy's Lizard file-nloc-medium gate (was 534 NLOC).

Grok Build: Grok 4.5 (high)
Linear: RM-193
